Move contents of Concludable to a new tool, ConstraintCopier
