+++ logic/concludable/Concludable.java
@@ -48,105 +48,4 @@ public abstract class Concludable<C extends Constraint, T extends Concludable<C,
         return constraint;
     }
 
-    static RelationConstraint copyConstraint(RelationConstraint relationConstraint) {
-        ThingVariable ownerCopy = copyIsaAndValues(relationConstraint.owner());
-        List<RelationConstraint.RolePlayer> rolePlayersCopy = copyRolePlayers(relationConstraint.players());
-        return new RelationConstraint(ownerCopy, rolePlayersCopy);
-    }
-
-    static List<RelationConstraint.RolePlayer> copyRolePlayers(List<RelationConstraint.RolePlayer> players) {
-        return players.stream().map(rolePlayer -> {
-            TypeVariable roleTypeCopy = rolePlayer.roleType().isPresent() ? copyVariableWithLabelAndValueType(rolePlayer.roleType().get()) : null;
-            ThingVariable playerCopy = copyIsaAndValues(rolePlayer.player());
-            RelationConstraint.RolePlayer rolePlayerCopy = new RelationConstraint.RolePlayer(roleTypeCopy, playerCopy);
-            rolePlayerCopy.addRoleTypeHints(rolePlayer.roleTypeHints());
-            return rolePlayerCopy;
-        }).collect(Collectors.toList());
-    }
-
-    static HasConstraint copyConstraint(HasConstraint hasConstraint) {
-        ThingVariable ownerCopy = copyIsaAndValues(hasConstraint.owner());
-        ThingVariable attributeCopy = copyIsaAndValues(hasConstraint.attribute());
-        return ownerCopy.has(attributeCopy);
-    }
-
-    static IsaConstraint copyConstraint(IsaConstraint isa) {
-        ThingVariable newOwner = ThingVariable.of(isa.owner().identifier());
-        copyValuesOntoVariable(isa.owner().value(), newOwner);
-        return copyIsaOntoVariable(isa, newOwner);
-    }
-
-    static ValueConstraint<?> copyConstraint(ValueConstraint<?> value) {
-        //NOTE: isa can never exist on a Value Concludable (or else it would be a Isa Concludable).
-        ThingVariable newOwner = ThingVariable.of(value.owner().identifier());
-        Set<ValueConstraint<?>> otherValues = value.owner().value().stream()
-                .filter(value1 -> !value.equals(value1)).collect(Collectors.toSet());
-        copyValuesOntoVariable(otherValues, newOwner);
-        return copyValueOntoVariable(value, newOwner);
-    }
-
-    static IsaConstraint copyIsaOntoVariable(IsaConstraint toCopy, ThingVariable variableToConstrain) {
-        TypeVariable typeCopy = copyVariableWithLabelAndValueType(toCopy.type());
-        IsaConstraint newIsa = variableToConstrain.isa(typeCopy, toCopy.isExplicit());
-        return newIsa;
-    }
-
-    static void copyValuesOntoVariable(Set<ValueConstraint<?>> toCopy, ThingVariable newOwner) {
-        toCopy.forEach(valueConstraint -> copyValueOntoVariable(valueConstraint, newOwner));
-    }
-
-    static ValueConstraint<?> copyValueOntoVariable(ValueConstraint<?> toCopy, ThingVariable toConstrain) {
-        if (toCopy.isLong())
-            return toConstrain.valueLong(toCopy.asLong().predicate(), toCopy.asLong().value());
-        else if (toCopy.isDouble())
-            return toConstrain.valueDouble(toCopy.asDouble().predicate(), toCopy.asDouble().value());
-        else if (toCopy.isBoolean())
-            return toConstrain.valueBoolean(toCopy.asBoolean().predicate(), toCopy.asBoolean().value());
-        else if (toCopy.isString())
-            return toConstrain.valueString(toCopy.asString().predicate(), toCopy.asString().value());
-        else if (toCopy.isDateTime())
-            return toConstrain.valueDateTime(toCopy.asDateTime().predicate(), toCopy.asDateTime().value());
-        else if (toCopy.isVariable()) {
-            ThingVariable copyOfVar = copyIsaAndValues(toCopy.asVariable().value());
-            return toConstrain.valueVariable(toCopy.asValue().predicate().asEquality(), copyOfVar);
-        } else throw GraknException.of(ILLEGAL_STATE);
-    }
-
-    static ThingVariable copyIsaAndValues(ThingVariable copyFrom) {
-        ThingVariable copy = ThingVariable.of(copyFrom.identifier());
-        copyIsaAndValues(copyFrom, copy);
-        return copy;
-    }
-
-    static void copyIsaAndValues(ThingVariable oldOwner, ThingVariable newOwner) {
-        if (oldOwner.isa().isPresent()) copyIsaOntoVariable(oldOwner.isa().get(), newOwner);
-        copyValuesOntoVariable(oldOwner.value(), newOwner);
-    }
-
-    static void copyLabelSubAndValueType(TypeVariable copyFrom, TypeVariable copyTo) {
-        if (copyFrom.label().isPresent()) copyTo.label(copyFrom.label().get().properLabel());
-        if (copyFrom.sub().isPresent()) {
-            SubConstraint subCopy = copyFrom.sub().get();
-            copyTo.sub(subCopy.type(), subCopy.isExplicit());
-        }
-        if (copyFrom.valueType().isPresent()) copyTo.valueType(copyFrom.valueType().get().valueType());
-    }
-
-    static TypeVariable copyVariableWithLabelAndValueType(TypeVariable copyFrom) {
-        TypeVariable copy = TypeVariable.of(copyFrom.identifier());
-        copyLabelSubAndValueType(copyFrom, copy);
-        return copy;
-    }
-
-    static boolean hasNoHints(Variable variable) {
-        return variable.isSatisfiable() && variable.resolvedTypes().isEmpty();
-    }
-
-    static boolean varHintsDisjoint(Variable conjVar, Variable thenVar) {
-        if (hasNoHints(conjVar) || hasNoHints(thenVar)) return false;
-        assert (conjVar.isThing() && thenVar.isThing()) ||
-                (conjVar.isType() && thenVar.isType());
-        return Collections.disjoint(conjVar.resolvedTypes(), thenVar.resolvedTypes());
-    }
-
 }
